rename the folder to v1
