+++ ui-spa/src/components/case-management/transfer-materials-v1/NavigationTable.module.scss
