+++ ui-spa/src/components/case-management/transfer-materials-v1/utils/isEgressFolderData.tsx
