+++ ui-spa/src/mocks/data/transferStatus.dev.ts
@@ -1,16 +1,33 @@
 import { type TransferStatusResponse } from "../../schemas";
 export const egressToNetAppTransferStatusDev: TransferStatusResponse = {
   id: "00000000-0000-4000-8000-000000000001",
-  status: "Completed",
+  status: "PartiallyCompleted",
   transferType: "Copy",
   direction: "EgressToNetApp",
   startedAt: null,
   completedAt: null,
-  failedItems: [],
+  failedItems: [
+    {
+      sourcePath: "folder1/folder2/file1.txt",
+      errorCode: "FileExists",
+    },
+    {
+      sourcePath: "folder1/folder2/file2.txt",
+      errorCode: "FileExists",
+    },
+    {
+      sourcePath: "folder1/folder3/file2.txt",
+      errorCode: "FileExists",
+    },
+    {
+      sourcePath: "folder1/folder2/file2.txt",
+      errorCode: "GeneralError",
+    },
+  ],
   userName: "dev_user@example.org",
   totalFiles: 30,
-  processedFiles: 30,
-  successfulFiles: 30,
+  processedFiles: 28,
+  successfulFiles: 28,
   failedFiles: 0,
   successItems: [
     {
